+++ libhesai/UdpParserGpu/src/udp_p64_parser_gpu.cu
@@ -187,7 +187,7 @@ int UdpP64ParserGpu<T_Point>::LoadCorrectionFile(std::string lidar_correction_fi
     LogDebug("Open correction file success");
     int length = 0;
     fin.seekg(0, std::ios::end);
-    length = fin.tellg();
+    length = static_cast<int>(fin.tellg());
     fin.seekg(0, std::ios::beg);
     char *buffer = new char[length];
     fin.read(buffer, length);
