+++ libhesai/UdpProtocol/udp_protocol_v3_2.h
@@ -249,7 +249,7 @@ struct HS_LIDAR_HEADER_QT_V2 {
 
   uint8_t GetLaserNum() const { return m_u8LaserNum; }
   uint8_t GetBlockNum() const { return m_u8BlockNum; }
-  double GetDistUnit() const { return m_u8DistUnit / 1000.f; }
+  float GetDistUnit() const { return m_u8DistUnit / 1000.f; }
   uint8_t GetEchoCount() const { return m_u8EchoCount; }
   bool IsFirstBlockLastReturn() const {
     return m_u8EchoCount == kFirstBlockLastReturn;
