+++ libhesai/UdpProtocol/udp_protocol_v4_3.h
@@ -239,7 +239,7 @@ struct HS_LIDAR_HEADER_ST_V3 {
 
   uint8_t GetLaserNum() const { return m_u8LaserNum; }
   uint8_t GetBlockNum() const { return m_u8BlockNum; }
-  double GetDistUnit() const { return m_u8DistUnit / 1000.f; }
+  float GetDistUnit() const { return m_u8DistUnit / 1000.f; }
   uint8_t GetEchoCount() const { return m_u8EchoCount; }
 
   bool IsFirstBlockLastReturn() const {
